Add a "plugin tweaks" module, and add a tweak for Internal Notes
The main mu-plugins directory on wporg is cluttered with files that are
for one or more "tweaks" to a particular plugin. The goal with this PR
is to not proliferate the same pattern in our pub-sync mu-plugins by
stashing all such tweaks in a "plugin-tweaks" subdirectory. Eventually
we could also clean up the main mu-plugins directory by migrating
open-sourcable tweaks to this new place.

The first tweak for this new system is for the Internal Notes plugin. It
started out as [a
PR](https://github.com/WordPress/pattern-directory/pull/463) for the
pattern directory, but after some discussion it seems better to put it
in mu-plugins since it will benefit any use of the Internal Notes plugin
anywhere on wporg.

Note that with this PR I haven't moved the existing "stream-tweaks" in
this repo into "plugin tweaks". This is because that particular tweak is
no longer needed, and will be removed in a separate PR.