Navigation menu: Fix compatibility with Gutenberg 16.0
In Gutenberg 16.0, the Navigation menu block changed how the classes are applied, so the custom `global-header__navigation` class is applied twice (on the nav and on the ul). The custom code to prevent menu wrapping is applied to both elements, but only removed from the nav, which caused the menu dropdowns to clip and not appear. Additionally, on block-theme sites, the font-size needs to be reset to inherit the header default.

The Navigation block also no longer uses micromodal, so the search button selectors need to be more general to work with both versions of Gutenberg.

Lastly, the Interactivity API expects a menu to have a child with class `wp-block-navigation-item`, and throws an error when it doesn't, so this simply adds the class to the search block.

See #405, https://meta.trac.wordpress.org/ticket/7068