Navigation: Prevent hiding the global header when navigation menus are open
The navigation JS for the global header previously applied a class when open, which let the header be hidden for non-header menus were open, which prevented an issue where the header overlapped the menu & close button on local navigation. Since the JS powering the navigation menu changed to the interactivity API, this global header JS was not applying anymore, and caused the header to disappear when the menu is clicked.

This removes that style, and adds a workaround for the local navigation menu.

Fixes #416