Remove focusout behavior
This was added to support the "click outside to close" behavior, but that can glitch when the click target is not a focusable element — clicking the space between buttons in the actions bar would close the dropdown. So the hasHover check was added to prevent that, but that brought in new glitches, where the focus could be reset to the first focusable element if hover was changed.

This PR removes all that in favor of using the backdrop element as a click target.