Block: Local navigation: Fixes for menu alignment & position.
Change positioning of local nav container when unsticky-ing it, use `relative` so that the brush stroke position works.
Fix alignment of menu on mid-sized screens— previous breakpoint was incorrect, the local nav menu does not switch to the responsive mode until 600px.
Remove padding on navigation menus with a background color, which lets menus set the color used in the modal background.

Fixes #445