New block: Query filter (#441)
* Add query filter block

* Refactor filters block

* Update styles

* Add action to inject extra field values

* Rename filter & improve documentation

* Update filter destination to use value from settings

* Rename variable

* Remove color configuration in favor of theme.json settings

* Create and use more custom theme.json values

* Set some defaults for custom properties

* Clear the form using the "clear" button

* Update the bubble as items are selected, but clear changes when closed without applying

* Use the `closeDropdown` helper to clear the form

* Add mobile styles

* Add key as an expected attribute

* Remove focusout behavior

This was added to support the "click outside to close" behavior, but that can glitch when the click target is not a focusable element — clicking the space between buttons in the actions bar would close the dropdown. So the hasHover check was added to prevent that, but that brought in new glitches, where the focus could be reset to the first focusable element if hover was changed.

This PR removes all that in favor of using the backdrop element as a click target.

* Remove unused class

* Fix spacing on mobile modal header

* Move keydown listener to parent container to catch focus on toggle button

* Check for span before updating the text

* Wrap options in a fieldset to associate collective title

This also adds a new label "title" which is used on the small screen modal too.