Google Map: Combine markers in the same location for better UX
If we did nothing, then the markers would overlap and only 1 would be accessible. Spreading the markers out is a common solution, but then the user would have to click on each one to know which ones are relevant to them. In many cases, it's a recurring event, so they really only need to know about the next one.

This way the user only has to click once, and they can see all of the info they need.

See https://ux.stackexchange.com/a/112281/13828