Plugin Compat: Fix Gutenberg loading for Blocks Everywhere
The version of Blocks Everywhere loaded on wporg was moved to `blocks-everywhere-1.23`, which broke this version loader. The main Gutenberg version was pinned to 18.0.1, which must be close enough to the supported 17.9 to work. Once unpinned, there are style issues (at least), so this will pin Blocks Everywhere back to the version it supports.