Local navigation bar block: Remove default background color property.
The custom color is set using `style.color.background`, not `backgroundColor`. So when using a custom color, technically there is no `backgroundColor` attribute. The change to correctly inherit default attribute values was then setting a backgroundColor, which the color block-support prioritizes when building the style code, overriding the custom color.

This default can be removed, the CSS applies the blueberry color if no other colors are set.

Fixes #654